# perf(stdlib): avoid using ValueForRow when appending values in window