- [x] docs/SPEC.md updated
- [x] Test cases written

It was found on a flamegraph that using `ValueForRow` and then appending
the value immediately generated a lot of allocations because it would
allocate memory for an interface and then discard it immediately.