@jsternberg 2019-05-03
Yea, it's kind of funky. I'm not sure why the number of allocations doesn't change.