@jsternberg 2019-05-03
It might be the time period. When I set the benchmark to generate points for 24 hours, I get these:

```
# before
jsternberg:[~/g/s/g/i/flux] $ (perf/append-row-value *+$=) go test -run=NONE -bench=FixedWindow_Process ./stdlib/universe/
goos: darwin
goarch: amd64
pkg: github.com/influxdata/flux/stdlib/universe
BenchmarkFixedWindow_Process-12                1        1848452786 ns/op        851720648 B/op  16008028 allocs/op
PASS
ok      github.com/influxdata/flux/stdlib/universe      3.208s
# after
jsternberg:[~/g/s/g/i/flux] $ (perf/append-row-value *$=) go test -run=NONE -bench=FixedWindow_Process ./stdlib/universe/
goos: darwin
goarch: amd64
pkg: github.com/influxdata/flux/stdlib/universe
BenchmarkFixedWindow_Process-12                1        1476167638 ns/op        663714152 B/op   7713627 allocs/op
PASS
ok      github.com/influxdata/flux/stdlib/universe      2.860s
```

When I do the benchmark in this commit, it's an 8 byte difference for allocations that seems to go out of control when dealing with larger tables. I'm going to see if I can improve the benchmark...